Use the new INotifyProcess1800 API from debugger (#9763)
* Add Microsoft.VisualStudio.Debugger.Interop.18.0 package references and enhance debugging interfaces

* Enhance OnAfterLaunchAsync method to include debug launch settings and improve Hot Reload support checks

* Refactor LaunchProfilesDebugLaunchProvider and ProjectLaunchTargetsProvider for improved process management and Hot Reload state handling

* Refactor Hot Reload session management and update interfaces for improved debugging support

* Add IDebugProfileLaunchTargetsProvider5 interface for enhanced launch process handling and remove unused code from ProjectLaunchTargetsProvider

* Refactor OnAfterLaunchAsync method to return completed task for improved performance

* Refactor LaunchProfilesDebugLaunchProvider and ProjectLaunchTargetsProvider to improve code clarity and remove unused dependencies

* Refactor IDebugProfileLaunchTargetsProvider5 interface and ProjectHotReloadSessionManager class to improve clarity and remove unused dependencies

* Refactor IProjectHotReloadSessionManager and ProjectHotReloadSessionManager to mark obsolete methods and improve session management; add unit tests for Hot Reload session functionality

* Refactor IDebugProfileLaunchTargetsProvider5 and related classes to improve documentation and remove obsolete methods; update session state handling in ProjectHotReloadSessionManager and enhance unit tests for clarity and maintainability

* Refactor IDebugProfileLaunchTargetsProvider5 documentation to simplify method references and update Public API declaration for OnAfterLaunchAsync

* Refactor ProjectHotReloadSessionManager to improve session state handling by optimizing the removal of active session states and updating the hot reload state notification logic.

* Refactor IDebugProfileLaunchTargetsProvider5 and LaunchProfilesDebugLaunchProvider to correct documentation and improve session handling; update IProjectHotReloadSessionManagerFactory for method signature consistency.
