+++ tests/Microsoft.VisualStudio.ProjectSystem.Managed.VS.UnitTests/Mocks/IProjectHotReloadSessionManagerFactory.cs
@@ -1,7 +1,10 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements. The .NET Foundation licenses this file to you under the MIT license. See the LICENSE.md file in the project root for more information.
 
+using Microsoft.VisualStudio.Debugger.Interop;
 using Microsoft.VisualStudio.ProjectSystem.Debug;
 using Microsoft.VisualStudio.ProjectSystem.HotReload;
+using Microsoft.VisualStudio.ProjectSystem.VS.HotReload;
+using Microsoft.VisualStudio.Shell.Interop;
 
 namespace Microsoft.VisualStudio.ProjectSystem.VS;
 
@@ -18,7 +21,7 @@ internal static class IProjectHotReloadSessionManagerFactory
             It.IsAny<ILaunchProfile>()))
             .ReturnsAsync(canCreatePendingSession);
 
-        mock.Setup(manager => manager.ActivateSessionAsync(It.IsAny<int>(), It.IsAny<string>()))
+        mock.Setup(manager => manager.ActivateSessionAsync(It.IsAny<IVsLaunchedProcess>(), It.IsAny<VsDebugTargetProcessInfo>()))
             .Returns(Task.CompletedTask);
 
         return mock.Object;
